feat(dashboard): file refactor
