+++ hooks/supabase/page/usePageQuery.tsx
@@ -1,6 +1,6 @@
 import { useQuery } from "react-query";
-import { supabase } from "../../utils/supabaseClient";
-import { definitions } from "../../types/supabase";
+import { supabase } from "../../../utils/supabaseClient";
+import { definitions } from "../../../types/supabase";
 import { pagesKey } from "./usePagesQuery";
 
 const getPage = async (pageId: string): Promise<definitions["pages"]> => {
