+++ hooks/supabase/post/useTogglePublishPostMutation.ts
@@ -1,6 +1,6 @@
 import { useMutation } from "react-query";
-import { supabase } from "../../utils/supabaseClient";
-import { definitions } from "../../types/supabase";
+import { supabase } from "../../../utils/supabaseClient";
+import { definitions } from "../../../types/supabase";
 
 const togglePublish = async (postId: number, publish: boolean) => {
   const { error } = await supabase
