+++ pages/dashboard/pages/index.tsx
@@ -3,7 +3,7 @@ import dynamic from "next/dynamic";
 import { useRequireAuth } from "../../../hooks/auth/useRequireAuth";
 import { useAuthUser } from "../../../hooks/auth/useAuthUser";
 import Dashboard from "../../../components/dashboard/Dashboard";
-import DashboardPages from "../../../components/dashboard/contents/DashboardPages";
+import DashboardPages from "../../../components/dashboard/contents/page/DashboardPages";
 
 const DashboardPagesPage = (): JSX.Element => {
   useRequireAuth();
