+++ pages/dashboard/posts/edit/[pid].tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { useRequireAuth } from "../../../../hooks/auth/useRequireAuth";
 import { useAuthUser } from "../../../../hooks/auth/useAuthUser";
-import PostComposer from "../../../../components/dashboard/contents/composer/PostComposer";
+import PostComposer from "../../../../components/dashboard/contents/post/PostComposer";
 import Dashboard from "../../../../components/dashboard/Dashboard";
 import dynamic from "next/dynamic";
 import { useRouter } from "next/router";
-import usePostQuery from "../../../../hooks/supabase/usePostQuery";
-import useUpdatePostMutation from "../../../../hooks/supabase/useUpdatePostMutation";
+import usePostQuery from "../../../../hooks/supabase/post/usePostQuery";
+import useUpdatePostMutation from "../../../../hooks/supabase/post/useUpdatePostMutation";
 import { definitions } from "../../../../types/supabase";
 import { toast } from "react-hot-toast";
